Fix "max_size exceeded in small_vector" error
This is a workaround for an error that occurs when there is a high
amount of packet loss and the client never clears its ACK packets.

The problem is in QuicClientTransport::processPacketData():
ackState.acks keeps growing because of non-consecutive packetNum entries
(due to packet loss). It eventually reaches a size above 16384 which is
the maximum value for uint16_t. This commit uses uint32_t instead, in
order to avoid this problem (or, at least, to delay its appearance).

Note: a better fix would ensure that ackStates.acks is cleared up over
time, instead of accumulating so much data!

+++ quic/codec/Types.h
@@ -44,7 +44,7 @@ constexpr auto kMaxPacketNumEncodingSize = 4;
 constexpr auto kNumInitialAckBlocksPerFrame = 32;
 
 template <class T>
-using IntervalSetVec = SmallVec<T, kNumInitialAckBlocksPerFrame, uint16_t>;
+using IntervalSetVec = SmallVec<T, kNumInitialAckBlocksPerFrame, uint32_t>;
 using AckBlocks = IntervalSet<PacketNum, 1, IntervalSetVec>;
 
 struct PaddingFrame {
